🚚 Move ListFilterBar component to /documents/components
